ai-service/requirements.txt
# AI Service Dependencies
supabase>=2.0.3
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

# AI Providers (install at least one)
google-generativeai>=0.3.0  # For Gemini
openai>=1.0.0               # For GPT
